Only set player weather if a plugin requested it.
If the server changes the weather it will set the per-player weather
variable and future changes will not apply. We should only set this
variable when a plugin is requesting per-player weather and not when
the server it doing it.

Remapped by Srg2Source from Bukkit/CraftBukkit@d7fe50c03da94ed5a3318cdc3a9a77c9c6afde26