Special case large chests for hopper events. Fixes BUKKIT-3916
Large chests work in a different fashion as they are a combination of
two other inventories. This causes their getOwner method to always return
null as their is no correct return. To compensate for this for the hopper
events we special case them to use their CraftBukkit counterpart that has
the information we need for the event.

Remapped by Srg2Source from Bukkit/CraftBukkit@7710efc5f925da754982dce263d6869be3f737b0